refactor(server): extract logging and error handling middleware
